requirements.txt
# Behavioral Health Platform - Core Dependencies
# Python >= 3.10

# Web Framework
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
starlette>=0.50.0

# Data Validation & Schema
pydantic>=2.12.5
pydantic-core>=2.41.5
email-validator>=2.1.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Clients
httpx>=0.28.1
aiohttp>=3.13.3
requests>=2.32.5

# Database & ORM
sqlalchemy>=2.0.45
aiosqlite>=0.22.1

# LLM & AI Integration
ollama>=0.6.1
openai>=2.15.0
llama-index>=0.14.12
llama-index-core>=0.14.12
llama-index-embeddings-ollama>=0.8.6
llama-index-llms-ollama>=0.9.1

# CLI & Terminal
click>=8.3.1
rich>=13.7.0
colorama>=0.4.6

# Utilities
python-dotenv>=1.2.1
pyyaml>=6.0.3
jinja2>=3.1.6
tenacity>=9.1.2

# NLP & Processing
nltk>=3.9.2
beautifulsoup4>=4.14.3

# Data Processing
pandas>=2.3.3
numpy>=2.4.1

# Database Migrations
alembic>=1.15.0

# Error Tracking
sentry-sdk[fastapi]>=2.0.0

# Monitoring
prometheus-fastapi-instrumentator>=7.0.0
redis>=5.0.0

# Testing (optional, can move to requirements-dev.txt)
pytest>=9.0.2
pytest-asyncio>=1.3.0
pytest-cov>=6.0.0

# Document Processing
pypdf>=6.6.0
pillow>=12.1.0

# Other
joblib>=1.5.3
tqdm>=4.67.1
typing-extensions>=4.15.0
